requirements.txt
azure-functions
python-dotenv
pydantic
